requirements.txt
accelerate==1.9.0
av==15.0.0
certifi==2025.8.3
charset-normalizer==3.4.2
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
hf-xet==1.1.5
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
imageio==2.37.0
jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lazy-loader==0.4
markupsafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnx==1.18.0
onnxruntime-gpu==1.22.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pip==25.2
protobuf==6.32.0rc1
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
qwen-vl-utils==0.0.11
regex==2025.7.34
requests==2.32.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.6.11
tokenizers==0.21.4
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.54.1
triton==3.3.1
typing-extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.175
ultralytics-thop==2.0.15
urllib3==2.5.0
